Shuffle column def code around
